+++ app/src/main/java/com/google/mediapipe/examples/gesturerecognizer/logic/ContextHolder.kt
@@ -1,10 +1,9 @@
 package com.google.mediapipe.examples.gesturerecognizer.logic
 
 import android.util.Log
-import com.google.mediapipe.tasks.components.containers.Category
-import com.google.mediapipe.tasks.components.containers.NormalizedLandmark
+import com.google.mediapipe.examples.gesturerecognizer.logic.model.GestureWrapper
+import com.google.mediapipe.examples.gesturerecognizer.logic.recognizer.impl.DownMovementRecognizer
 import com.google.mediapipe.tasks.vision.gesturerecognizer.GestureRecognizerResult
-import com.google.mediapipe.tasks.vision.handlandmarker.HandLandmark
 
 
 object ContextHolder {
@@ -15,7 +14,7 @@ object ContextHolder {
     private val dynamicSignCandidatesMap = mapOf(
         "A" to "Ą",
         "C" to "Ć",
-        "CZ" to "CH", // te znaki to weryfikacji
+        "CZ" to "CH",
         "D" to "D",
         "E" to "Ę",
         "F" to "F",
@@ -78,14 +77,8 @@ object ContextHolder {
         }
     }
 
-//    private fun checkHandMovement(
-//        first: List<NormalizedLandmark>, last: List<NormalizedLandmark>
-//    ): Float {
-//        return last[HandLandmark.WRIST].y() - first[HandLandmark.WRIST].y()
-//    }
-
     private fun matchDynamicGesture(label: String) {
-        Log.i(CONTEXT_HOLDER_TAG, "matching dynamic gesture\n Letter = $label")
+        Log.i(CONTEXT_HOLDER_TAG, "matching dynamic gesture\nLetter = $label")
         when (label) {
             "N", "O", "C" -> {
                 val movement = DownMovementRecognizer().checkHandMovement(gesturesArray)
@@ -119,60 +112,5 @@ object ContextHolder {
         if (gesturesArray.none { it.getCategory() == gesture.getCategory() }) {
             gesturesArray.clear()
         }
-    }
-}
-
-data class GestureWrapper(
-    val gesture: MutableList<MutableList<Category>>,
-    val landmarks: MutableList<MutableList<NormalizedLandmark>>
-) {
-    fun getCategory(): String {
-        return gesture[0][0].categoryName()
-    }
-
-    fun getLandmarkArray(): List<NormalizedLandmark> {
-        return landmarks[0]
-    }
-}
-
-interface DynamicGestureRecognizer {
-    fun checkHandMovement(gestureList: List<GestureWrapper>): Boolean
-}
-
-class DownMovementRecognizer : DynamicGestureRecognizer {
-    override fun checkHandMovement(gestureList: List<GestureWrapper>): Boolean {
-        for (gest in gestureList)
-            Log.i(TAG, "${gest.getLandmarkArray()}")
-        if (gestureList.size < MIN_LENGTH)
-            return false
-        val movements = mutableListOf<Float>()
-        for (i in gestureList.size - 1 downTo 1) {
-            // size = 4
-            // brane indeksy: 3,2; 2,1; 1,0
-            // porownuje (od dołu do gory) najnowszy z drugim najnowszym, w tym przypadku:
-            // roznica powinna byc > 0, bo najnowsza ma byc nizej niz druga najnowsza ->
-            val movement = comparePositions(gestureList[i], gestureList[i - 1])
-            Log.i("checkHandMovement", "$movement between indexed $i and ${i - 1}")
-            movements.add(movement)
-        }
-        return movements.all { it > MOVEMENT_THRESHOLD }
-    }
-
-    private fun comparePositions(
-        first: GestureWrapper, second: GestureWrapper
-    ): Float {
-        return first.getLandmarkArray()[HandLandmark.WRIST].y() - second.getLandmarkArray()[HandLandmark.WRIST].y()
-    }
-
-    companion object {
-        private const val MIN_LENGTH: Int = 4
-        private const val TAG: String = "DownMovementRecognizer"
-        private const val MOVEMENT_THRESHOLD: Float = 0.001F
-    }
-}
-
-class UpMovementRecognizer : DynamicGestureRecognizer {
-    override fun checkHandMovement(gestureList: List<GestureWrapper>): Boolean {
-        return true
     }
 }
